fix: 로또 1등 당청금 수정

+++ src/main/java/lotto/modal/Money.java
@@ -20,7 +20,7 @@ public class Money {
 		return this.userMoney / LOTTO_PRICE_MONEY;
 	}
 
-	public BigDecimal getYield(int totalPrize) {
+	public BigDecimal getYield(long totalPrize) {
 
 		BigDecimal money = new BigDecimal(this.userMoney);
 		BigDecimal prize = new BigDecimal(totalPrize);
